+++ Source/MintegralAdapter.swift
@@ -149,7 +149,12 @@ final class MintegralAdapter: PartnerAdapter {
         case .banner:
             return try MintegralAdapterBannerAd(adapter: self, request: request, delegate: delegate)
         default:
-            throw error(.loadFailureUnsupportedAdFormat)
+            // Not using the `.adaptiveBanner` case directly to maintain backward compatibility with Chartboost Mediation 4.0
+            if request.format.rawValue == "adaptive_banner" {
+                return try MintegralAdapterBannerAd(adapter: self, request: request, delegate: delegate)
+            } else {
+                throw error(.loadFailureUnsupportedAdFormat)
+            }
         }
     }